Move the megabytes convertion helper to the new view helper
